# ScorecardAPI/openapi.yaml
openapi: 3.0.1
info:
  title: ScorecardAPI
  version: "1.0"
servers:
- url: /
paths:
  /api/Matches:
    get:
      tags:
      - Matches
      operationId: GetMatches
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchOutputDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchOutputDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchOutputDTO'
    post:
      tags:
      - Matches
      operationId: CreateMatch
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CreateMatchDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMatchDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateMatchDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateMatchDTO'
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Match'
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
            text/json:
              schema:
                $ref: '#/components/schemas/Match'
  /api/Matches/{matchId}:
    get:
      tags:
      - Matches
      operationId: GetMatch
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MatchOutputDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/MatchOutputDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/MatchOutputDTO'
    put:
      tags:
      - Matches
      operationId: ModifyMatch
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ModifyMatchDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyMatchDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/ModifyMatchDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ModifyMatchDTO'
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Match'
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
            text/json:
              schema:
                $ref: '#/components/schemas/Match'
    delete:
      tags:
      - Matches
      operationId: DeleteMatch
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
  /api/Matches/{matchId}/status:
    put:
      tags:
      - Matches
      operationId: SetStatus
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/MatchStatus'
          application/json:
            schema:
              $ref: '#/components/schemas/MatchStatus'
          text/json:
            schema:
              $ref: '#/components/schemas/MatchStatus'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MatchStatus'
      responses:
        "200":
          description: Success
  /api/Matches/{matchId}/events:
    get:
      tags:
      - Matches
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MatchEventOutputDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/MatchEventOutputDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/MatchEventOutputDTO'
  /api/Matches/{matchId}/events/strike:
    post:
      tags:
      - Matches
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/StrikeEventDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/StrikeEventDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/StrikeEventDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StrikeEventDTO'
      responses:
        "200":
          description: Success
  /api/Matches/{matchId}/events/penalty:
    post:
      tags:
      - Matches
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PenaltyEventDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/PenaltyEventDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/PenaltyEventDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PenaltyEventDTO'
      responses:
        "200":
          description: Success
  /api/Matches/{matchId}/events/disqualify:
    post:
      tags:
      - Matches
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/DisqualificationEventDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/DisqualificationEventDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/DisqualificationEventDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DisqualificationEventDTO'
      responses:
        "200":
          description: Success
  /api/Matches/{matchId}/events/{eventId}:
    delete:
      tags:
      - Matches
      operationId: DeleteEvent
      parameters:
      - name: matchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      - name: eventId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
...
components:
  schemas:
    CreateMatchDTO:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int32
        fighterOneId:
          type: integer
          format: int32
        fighterTwoId:
          type: integer
          format: int32
        duration:
          type: integer
          format: int32
        pool:
          type: integer
          format: int32
      additionalProperties: false
...
    DisqualificationEventDTO:
      type: object
      properties:
        fighterOneReduction:
          type: integer
          format: int32
        fighterTwoReduction:
          type: integer
          format: int32
        applyHealthReduction:
          type: boolean
        userId:
          type: integer
          format: int32
        reason:
          type: string
          nullable: true
      additionalProperties: false
    Match:
      type: object
      properties:
        matchId:
          type: integer
          format: int32
        tournamentId:
          type: integer
          format: int32
        fighterOneId:
          type: integer
          format: int32
        fighterTwoId:
          type: integer
          format: int32
        fighterOneHealth:
          type: integer
          format: int32
        fighterTwoHealth:
          type: integer
          format: int32
        pool:
          type: integer
          format: int32
        duration:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/MatchStatus'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        events:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/MatchEvent'
        tournament:
          $ref: '#/components/schemas/Tournament'
      additionalProperties: false
    MatchEvent:
      type: object
      properties:
        matchEventId:
          type: integer
          format: int32
        matchId:
          type: integer
          format: int32
        match:
          $ref: '#/components/schemas/Match'
        applyHealthReduction:
          type: boolean
        eventType:
          type: string
          nullable: true
        eventTime:
          type: string
          format: date-time
        fighterOneReduction:
          type: integer
          format: int32
        fighterTwoReduction:
          type: integer
          format: int32
      additionalProperties: false
    MatchEventOutputDTO:
      type: object
      properties:
        matchId:
          type: integer
          format: int32
        eventId:
          type: integer
          format: int32
        eventType:
          type: string
          nullable: true
        eventTime:
          type: string
          nullable: true
        fighterOneReduction:
          type: integer
          format: int32
        fighterTwoReduction:
          type: integer
          format: int32
      additionalProperties: false
    MatchOutputDTO:
      type: object
      properties:
        matchId:
          type: integer
          format: int32
        tournamentId:
          type: integer
          format: int32
        fighterOneId:
          type: integer
          format: int32
        fighterTwoId:
          type: integer
          format: int32
        fighterOneHealth:
          type: integer
          format: int32
        fighterTwoHealth:
          type: integer
          format: int32
        pool:
          type: integer
          format: int32
        duration:
          type: integer
          format: int32
        started:
          type: string
          nullable: true
        ended:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/MatchStatus'
      additionalProperties: false
    MatchStatus:
      type: integer
      format: int32
      enum:
      - 0
      - 1
      - 2
      - 3
    ModifyMatchDTO:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int32
        fighterOneId:
          type: integer
          format: int32
        fighterTwoId:
          type: integer
          format: int32
        fighterOneHealth:
          type: integer
          format: int32
        fighterTwoHealth:
          type: integer
          format: int32
        duration:
          type: integer
          format: int32
      additionalProperties: false
...
    PenaltyEventDTO:
      type: object
      properties:
        fighterOneReduction:
          type: integer
          format: int32
        fighterTwoReduction:
          type: integer
          format: int32
        applyHealthReduction:
          type: boolean
        reason:
          type: string
          nullable: true
      additionalProperties: false
    StrikeEventDTO:
      type: object
      properties:
        fighterOneReduction:
          type: integer
          format: int32
        fighterTwoReduction:
          type: integer
          format: int32
        applyHealthReduction:
          type: boolean
      additionalProperties: false
    Tournament:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int32
        matches:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Match'
        fighters:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/User'
      additionalProperties: false
...
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        tournaments:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Tournament'
      additionalProperties: false
...